Fixing multiple instances bug (each instance has the same username/pass)
Apparently Reslter's service functionality cannot handle multiple instances correctly.

In Restler the only way to set the username/password for the service is in the defaults.
But the defaults is a static property so it's shared across all instances of the service.

Guess it's time to move away from Restler. Best alternative?